# backend/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Ecommerce",
    "description": "API du site d'ecommerce WeThePast",
    "version": "1.0.0"
  },
  "host": "localhost:5000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/api/shoes/seed": {
      "get": {
        "description": "seed",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/shoes/": {
      "get": {
        "description": "renvoie toutes les chaussures du site",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/shoes/search/{searchTerm}": {
      "get": {
        "description": "permet de renvoyer les chaussures qui correspondent à un terme de recherche",
        "parameters": [
          {
            "name": "searchTerm",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/shoes/tags": {
      "get": {
        "description": "renvoie les tags qui correspondent à une chaussure",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/shoes/tag/{tagName}": {
      "get": {
        "description": "renvoie les chaussures qui correspondent a un tag donné",
        "parameters": [
          {
            "name": "tagName",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/shoes/{shoesId}": {
      "get": {
        "description": "permet de renvoyer un chaussure via son ID",
        "parameters": [
          {
            "name": "shoesId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/users/seed": {
      "get": {
        "description": "seed",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/users/login": {
      "post": {
        "description": "permet de créer un nouveau profil de login",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/users/register": {
      "post": {
        "description": "enregistre un nouveau profil dans la db en cryptant le password",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                },
                "address": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/orders/create": {
      "post": {
        "description": "test",
        "parameters": [
          {
            "name": "access_token",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "items": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/orders/newOrderForCurrentUser": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "access_token",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/orders/pay": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "access_token",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "paymentId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  }
}